NCBC-3596: Support for maxTTL value of -1 for collection "no expiry"
Motivation
----------
Allows to set no document expiry on a collection, rather than using the bucket's expiry.
This can already be done in C# by passing a TimeSpan.FromSeconds(-1)
Changes
-------
- Updated code documentation to specify supported values for MaxExpiry in Create/UpdateCollectionSettings
- Updated BucketSettings code documentation to do the same.
- Fixed bug where the MaxExpiry Timespan in the UpdateCollection method wasn't using its TotalSeconds

Reviewed-on: https://review.couchbase.org/c/couchbase-net-client/+/203498
Tested-by: Build Bot <build@couchbase.com>

+++ src/Couchbase/Management/Collections/CollectionManager.cs
@@ -235,7 +235,7 @@ namespace Couchbase.Management.Collections
 
                 if (settings.MaxExpiry.HasValue)
                 {
-                    keys.Add("maxTTL", settings.MaxExpiry.Value.TotalSeconds.ToString(CultureInfo.InvariantCulture));
+                    keys.Add("maxTTL", ((int)settings.MaxExpiry.Value.TotalSeconds).ToString(CultureInfo.InvariantCulture));
                 }
                 if (settings.History.HasValue)
                 {
@@ -499,7 +499,7 @@ namespace Couchbase.Management.Collections
 
             if (settings.MaxExpiry.HasValue)
             {
-                dict.Add("maxTTL", settings.MaxExpiry.Value.TotalSeconds.ToString(CultureInfo.InvariantCulture));
+                dict.Add("maxTTL", ((int)settings.MaxExpiry.Value.TotalSeconds).ToString(CultureInfo.InvariantCulture));
             }
 
             if (settings.History.HasValue)
